SIIT/NAT64 kernel/userspace binaries look stable.
The graybox tests are passing in the development kernels.

Still missing:

- Review of joold.
- Merge issue189.
- Merge master.

+++ include/nat64/usr/global.h
@@ -46,7 +46,7 @@
 #define OPTNAME_SYNCH_PERIOD		"synch-period"
 
 int global_display(bool csv);
-int global_update(__u8 type, size_t size, void *data);
+int global_update(__u16 type, size_t size, void *data);
 
 
 #endif /* _JOOL_USR_GLOBAL_H */
